refactor: remoce parenthesis from home folder and update related paths
