+++ app/page.tsx
@@ -1,6 +1,6 @@
 import { JSX } from 'react';
 import './globals.css';
-import Home from './(home)/Home';
+import Home from './home/Home';
 
 const Page = (): JSX.Element => {
   return (
